+++ app/src/main/java/kids/baba/mobile/core/utils/EncryptedPrefs.kt
@@ -14,7 +14,7 @@ object EncryptedPrefs {
 
     fun initSharedPreferences(context: Context) {
         synchronized(this) {
-            val sharedPrefsFile = "BaBaPrefs2"
+            val sharedPrefsFile = "BaBaPrefs"
             val masterKey = MasterKey.Builder(context)
                 .setKeyScheme(MasterKey.KeyScheme.AES256_GCM)
                 .build()
