+++ app/src/main/java/kids/baba/mobile/data/api/AlbumApi.kt
@@ -3,14 +3,7 @@ package kids.baba.mobile.data.api
 import kids.baba.mobile.domain.model.*
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
-import retrofit2.Response
-import retrofit2.http.GET
-import retrofit2.http.Header
-import retrofit2.http.Multipart
-import retrofit2.http.POST
-import retrofit2.http.Part
-import retrofit2.http.PartMap
-import retrofit2.http.Path
+import retrofit2.http.*
 
 interface AlbumApi {
 
@@ -25,9 +18,9 @@ interface AlbumApi {
         @Header("Authorization")
         accessToken: String,
         @Path("babyId") id: String,
-        @Part photo: MultipartBody.Part?,
+        @Part photo: MultipartBody.Part,
         @PartMap bodyDataHashMap: HashMap<String, RequestBody>
-        ): PostAlbumResponse/*: Response<Article>*/
+    ): PostAlbumResponse
 
 
     //성장 앨범 좋아요
