--- app/src/main/java/kids/baba/mobile/domain/model/ArticleResponse.kt
@@ -1,5 +1,0 @@
-package kids.baba.mobile.domain.model
-
-data class ArticleResponse(
-    val code: String
-)
