+++ app/src/main/java/kids/baba/mobile/presentation/view/GrowthAlbumFragment.kt
@@ -334,7 +334,7 @@ class GrowthAlbumFragment : Fragment() {
         // 카메라 권한을 사용하는 무엇이든지 호출하면 이 메서드 실행
         Log.d("GrowthAlbumFragment", "connect()")
         // TODO: 여기서 nowDate 을 String 으로 넘겨주어야 합니다.
-        FilmActivity.startActivity(requireContext(), nowDate = "2023-02-04")
+        FilmActivity.startActivity(requireContext(), nowDate = "2023-04-01")
     }
 
 }
