+++ app/src/main/java/kids/baba/mobile/presentation/viewmodel/CropViewModel.kt
@@ -1,14 +1,11 @@
 package kids.baba.mobile.presentation.viewmodel
 
-import android.content.Context
-import android.os.Environment
 import android.util.Log
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.canhub.cropper.CropImageView
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dagger.hilt.android.qualifiers.ApplicationContext
 import kids.baba.mobile.domain.model.MediaData
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.channels.trySendBlocking
@@ -21,13 +18,15 @@ import javax.inject.Inject
 @HiltViewModel
 class CropViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
-    @ApplicationContext private val context: Context
+    private val outputDirectory: File,
 ) : ViewModel() {
 
     private val TAG = "CropViewModel"
 
     val currentTakenMediaInCrop = savedStateHandle.get<MediaData>(MEDIA_DATA)
-    val currentTakenMedia = MutableStateFlow<MediaData?>(savedStateHandle[MEDIA_DATA])
+    private val currentTakenMedia =
+        MutableStateFlow(savedStateHandle[WriteTitleViewModel.MEDIA_DATA] ?: MediaData())
+
 
     fun cropImage(cropImageView: CropImageView) = callbackFlow {
         viewModelScope.launch {
@@ -38,8 +37,7 @@ class CropViewModel @Inject constructor(
                 )
 
                 val fileName = result.uriContent.toString().split("/").last()
-                val storageDir: File? = context.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-                val file = File(storageDir, fileName)
+                val file = File(outputDirectory, fileName)
 
                 currentTakenMedia.value =
                     MediaData(
@@ -47,7 +45,7 @@ class CropViewModel @Inject constructor(
                         mediaUri = file.absolutePath
                     )
 
-                trySendBlocking(currentTakenMedia.value!!)
+                trySendBlocking(currentTakenMedia.value)
             }
             cropImageView.croppedImageAsync()
